+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ Its primary motivation was to make the behavior of REST API v2 client objects mo
    - The ``api_key`` property has been removed from all client classes except ``pagerduty.RestApiV2BaseClient`` and subclasses thereof, wherein it is deprecated.
    - The hook method ``after_set_api_key`` will be ignored in all classes where the ``api_key`` property has been removed.
 
+**2025-09-08: Bugfix: iter_all parameter serialization - Version 4.1.1**
+
+* Fixes a regression wherein a boolean is serialized to the capitalized ``True`` / ``False`` for the ``total`` parameter in classic pagination, whereas the API requires lower case values.
+
 **2025-08-27: Feature: PKCE OAuth helpers - Version 4.1.0**
 
 * **Feature:** new methods ``get_new_token_from_code_with_pkce``, ``generate_s256_pkce_params`` and ``get_pkce_authorize_url`` of ``pagerduty.OAuthTokenClient`` to assist with implementation of the OAuth via PKCE token grant flow.